Extract MobileAlbumCard to shared component
- Move MobileAlbumCard + ArtistLinks from AlbumTable.jsx to MobileAlbumCard.jsx
- Import back into AlbumTable — no behavioral change
- Enables reuse in upcoming SearchOverlay component
- Update spec and plan for full-screen search overlay design
