+++ docs/specs/2026-04-16-mobile-search-ux-design.md
@@ -1,66 +1,120 @@
-# Mobile Search UX Fix — Design Spec
+# Mobile Search UX — Design Spec
 
 **Issue:** [#28](https://github.com/toofanian/bummer/issues/28)
-**Date:** 2026-04-16
+**Date:** 2026-04-16 (revised 2026-04-17)
 
 ## Problem
 
-Mobile search UX has three concrete issues:
-1. Header (with search input) scrolls away with content — not fixed/sticky
-2. BulkAddBar (z-50) renders behind BottomTabBar (z-200) — z-index stacking bug
-3. Search input has minimal focus feedback (border-color only)
+Search on mobile is an afterthought — a tiny inline input crammed into the header alongside title, toggle, and settings button. ~150px usable width. Only appears on library/collections views. Not discoverable, not comfortable to use. Search should be primary navigation, not a filter box.
 
-## Scope
+## Solution: Full-screen search overlay
 
-Three targeted fixes. No architectural changes.
+Replace inline search input with a search icon in the header that opens a full-screen search overlay. Follows the Spotify/Apple Music pattern.
 
-### Out of scope
-- Search state persistence across tab switches — intentionally clears on tab switch (existing behavior, confirmed desired)
-- iOS keyboard handling — `h-dvh` + `viewport-fit=cover` + `overflow-y-auto` already handle this correctly on modern iOS Safari
-- Desktop layout — unaffected, separate code path
+## Out of scope
+- Search persistence across dismiss (intentionally clears)
+- Desktop layout changes (separate code path)
+- Search across non-library content (collections search comes later if needed)
+- iOS keyboard handling (h-dvh + viewport-fit=cover already correct)
 
-## Fix 1: Sticky header
+---
 
-**File:** `frontend/src/App.jsx` (mobile layout, ~line 673)
+## Trigger: Search icon in header
 
-**Current:** Header is a regular flow element inside `flex flex-col h-dvh`. Content area is `flex-1 overflow-hidden` with inner `overflow-y-auto`.
+**File:** `frontend/src/App.jsx` (mobile header, ~line 717)
 
-**Change:** Add `sticky top-0 z-[100]` and explicit `bg-surface` to the header element. Since the header is already outside the scroll container (content scrolls inside `overflow-y-auto` div), this should work with the existing flex layout. The `z-[100]` ensures header stays above scrolling content but below fixed bottom bars (z-190+).
+- Remove inline `<input>` from header (lines 730-737)
+- Add search icon button in header, visible on **all views** (not just library/collections)
+- Icon: magnifying glass (inline SVG, match existing icon style)
+- Tapping opens SearchOverlay
+- New state: `const [searchOpen, setSearchOpen] = useState(false)`
 
-**Risk:** Low. Header is already outside scroll container. Sticky just prevents it from collapsing out of the flex layout on scroll.
+**Header after change:**
+```
+[Title] [LibraryViewToggle?] [SearchIcon] [SettingsIcon]
+```
 
-## Fix 2: BulkAddBar z-index
+## SearchOverlay component
 
-**File:** `frontend/src/components/BulkAddBar.jsx` (~line 12)
+**New file:** `frontend/src/components/SearchOverlay.jsx`
 
-**Current:** `z-50` (Tailwind z-50 = 50). BottomTabBar is `z-[200]`. BulkAddBar renders behind BottomTabBar.
+**Position & z-index:**
+- `fixed inset-0 z-[350]` — full-screen, above playback bars (z-190/200/300) but below modals (z-400+)
+- `bg-surface` — opaque, not transparent/blurred
+- Respects safe-area-inset-top and safe-area-inset-bottom
 
-**Change:** BulkAddBar should sit above BottomTabBar. Change to `z-[210]`. Also adjust `bottom` offset so BulkAddBar sits above BottomTabBar rather than overlapping at `bottom-0`:
-- Set `bottom` to `calc(50px + env(safe-area-inset-bottom, 0px))` (BottomTabBar height)
+**Layout (top to bottom):**
+1. **Search bar row** — `sticky top-0`, safe-area-inset-top padding
+   - Search input: full width, autofocused, large tap target (py-3), text-base (not text-sm)
+   - Cancel button: text button, right side, dismisses overlay
+2. **Results area** — `flex-1 overflow-y-auto`
+   - Renders `MobileAlbumCard` for each match (reuse from AlbumTable.jsx)
+   - Uses existing `filterAlbums()` logic against full album list
+   - Empty query: show nothing (blank) or optional "Start typing to search" hint
+   - No matches: "No results" text
+3. **No tab bar** — BottomTabBar hidden while overlay is open
 
-**Z-index ladder after fix:**
-- z-[100]: Sticky header
-- z-[190]: MiniPlaybackBar
-- z-[200]: BottomTabBar
-- z-[210]: BulkAddBar
-- z-[300]: DigestPanel, FullScreenNowPlaying
-- z-[400-401]: DevicePicker modal
+**Props:**
+```js
+{
+  albums,           // full album list to search
+  search,           // current search string
+  onSearchChange,   // setter
+  onClose,          // dismiss overlay
+  onPlay,           // play album/track
+  playback,         // current playback state (for now-playing indicator)
+  // pass through any props MobileAlbumCard needs
+}
+```
 
-## Fix 3: Search input focus states
+**Behavior:**
+- Input autofocuses on mount
+- ESC key or Cancel button dismisses
+- Search string clears on dismiss
+- Album cards behave identically to library view: expand tracks, tap to play
+- BulkAddBar still works during search (selectedAlbumIds state lives in App.jsx)
 
-**File:** `frontend/src/App.jsx` (~line 687) and/or `frontend/src/tailwind.css` (~line 99)
+## State wiring in App.jsx
 
-**Current:** Only `border-color: var(--color-focus-border)` on focus. Very subtle.
+- New state: `searchOpen` (boolean)
+- Existing `search` / `setSearch` state reused, wired to overlay instead of inline input
+- On overlay close: `setSearchOpen(false); setSearch('')`
+- BottomTabBar: conditionally hidden when `searchOpen` (or overlay covers it via z-index)
+- MiniPlaybackBar: stays visible below overlay? **No** — overlay is full-screen opaque, covers everything
 
-**Change:** Add visible focus ring and background shift:
-- `focus:ring-2 focus:ring-accent/40 focus:bg-surface-2/80` on the search input element
-- Or update the global `input:focus` rule in `tailwind.css` to include ring + background
+## MobileAlbumCard extraction
 
-Prefer input-specific classes over global rule change to avoid unintended side effects.
+`MobileAlbumCard` is currently defined inside `AlbumTable.jsx` (~line 72). Need to either:
+- **Option A:** Extract to own file `components/MobileAlbumCard.jsx` for reuse
+- **Option B:** Import/export from AlbumTable
+
+**Go with A** — cleaner, AlbumTable.jsx is already large.
+
+## Sticky header fix (from original spec)
+
+Keep `sticky top-0 z-[100]` on mobile header. Already applied.
+
+## Search input focus states
+
+Move to SearchOverlay input instead. Larger input + ring focus:
+- `focus:ring-2 focus:ring-accent/40 focus:outline-none`
+- Larger size: `text-base py-3 px-4` (not the old text-sm py-1)
+
+## Z-index ladder (final)
+
+| z-index | Component |
+|---------|-----------|
+| 100 | Sticky header |
+| 190 | MiniPlaybackBar |
+| 200 | BottomTabBar |
+| 300 | BulkAddBar |
+| 350 | **SearchOverlay** |
+| 400-401 | DevicePicker modal |
+| 500 | CollectionPicker modal |
 
 ## Testing
 
-- **Sticky header:** Scroll content on mobile viewport; header should remain visible at top
-- **BulkAddBar:** Enter bulk-add mode; bar should appear above BottomTabBar, not behind it
-- **Focus states:** Tap search input; should see visible ring/glow feedback
-- Unit tests: existing Vitest tests should still pass (no behavioral changes)
+- **Unit (Vitest):** SearchOverlay renders, input autofocuses, typing filters results, cancel dismisses, escape dismisses
+- **Unit:** Header shows search icon on all mobile views
+- **Unit:** MobileAlbumCard renders identically when used from SearchOverlay vs AlbumTable
+- **Existing tests:** AlbumTable tests still pass after MobileAlbumCard extraction
